requirements.txt
asgiref==3.2.10
Django==3.0.8
django-grpc==1.0.8
grpcio==1.30.0
grpcio-tools==1.30.0
protobuf==3.12.4
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
